third_party/blink/web_tests/external/wpt/css/css-inline/parsing/baseline-shift-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['baseline-shift'] = "0" should set the property value
  assert_equals: serialization should be canonical expected "0px" but got "0"
Harness: the test ran to completion.
